chore: add type-check script
