+++ package.json
@@ -14,6 +14,7 @@
     "build": "npm run build:client && npm run build:server",
     "build:dev": "vite build --mode development",
     "lint": "eslint .",
+    "type-check": "tsc --noEmit",
     "preview": "vite preview",
     "start": "node server/index.js"
   },
