+++ src/components/waiver/CustomerWaiverForm.tsx
@@ -45,7 +45,6 @@ const CustomerWaiverForm = ({ equipment, onComplete }: CustomerWaiverFormProps) 
     }
 
     // Simulating form submission
-    console.log("Form Data:", { ...data, signature });
     
     toast({
       title: "Waiver Submitted",
