+++ src/services/blogSearchService.ts
@@ -2,7 +2,6 @@
 import { BlogPost } from "@/lib/blog";
 
 export const searchBlogPostsWithNLP = async (query: string, posts: BlogPost[]): Promise<BlogPost[]> => {
-  console.log(`Processing blog search query: "${query}"`);
 
   // Convert query to lowercase for easier matching
   const lowerQuery = query.toLowerCase();
