+++ src/services/equipment/appSettingsService.ts
@@ -4,7 +4,6 @@ import { supabase } from "@/integrations/supabase/client";
 // Get global app setting for mock data
 export const getShowMockDataSetting = async (): Promise<boolean> => {
   try {
-    console.log('🔍 Checking global app setting for show_mock_data...');
     const { data, error } = await supabase
       .from('app_settings')
       .select('setting_value')
@@ -13,16 +12,13 @@ export const getShowMockDataSetting = async (): Promise<boolean> => {
 
     if (error) {
       console.error('❌ Error fetching app setting:', error);
-      console.log('🔄 Defaulting to mock data due to error');
       return true; // Default to mock data if error
     }
 
     const useMockData = data?.setting_value === true;
-    console.log('✅ Global app setting show_mock_data:', useMockData);
     return useMockData;
   } catch (error) {
     console.error('❌ Exception fetching app setting:', error);
-    console.log('🔄 Defaulting to mock data due to exception');
     return true; // Default to mock data if error
   }
 };
@@ -30,7 +26,6 @@ export const getShowMockDataSetting = async (): Promise<boolean> => {
 // Get global app setting for AI search usage
 export const getUseAISearchSetting = async (): Promise<boolean> => {
   try {
-    console.log('🔍 Checking global app setting for use_ai_search...');
     const { data, error } = await supabase
       .from('app_settings')
       .select('setting_value')
@@ -39,16 +34,13 @@ export const getUseAISearchSetting = async (): Promise<boolean> => {
 
     if (error) {
       console.error('❌ Error fetching app setting:', error);
-      console.log('🔄 Defaulting to AI search due to error');
       return true;
     }
 
     const useAI = data?.setting_value === true;
-    console.log('✅ Global app setting use_ai_search:', useAI);
     return useAI;
   } catch (error) {
     console.error('❌ Exception fetching app setting:', error);
-    console.log('🔄 Defaulting to AI search due to exception');
     return true;
   }
 };
