+++ src/services/equipmentUpdateService.ts
@@ -7,17 +7,11 @@ export const updateEquipmentInDatabase = async (
   equipmentData: Record<string, unknown>,
   userId: string
 ) => {
-  console.log('=== EQUIPMENT UPDATE SERVICE ===');
-  console.log('Equipment ID:', equipment.id);
-  console.log('Current User ID (editor):', userId);
-  console.log('Original Equipment Owner ID:', equipment.user_id);
-  console.log('Equipment data to update:', equipmentData);
 
   // CRITICAL: Never update user_id to preserve original ownership
   // Remove user_id from equipmentData if it exists to ensure we don't change ownership
   if ('user_id' in equipmentData) {
     delete equipmentData.user_id;
-    console.log('Removed user_id from update data to preserve original ownership');
   }
 
   // Validate that we have required data
@@ -31,7 +25,6 @@ export const updateEquipmentInDatabase = async (
 
   // Ensure damage_deposit is properly handled
   if (equipmentData.damage_deposit !== undefined) {
-    console.log('Damage deposit value:', equipmentData.damage_deposit, 'Type:', typeof equipmentData.damage_deposit);
     
     // Convert to number if it's a string
     if (typeof equipmentData.damage_deposit === 'string') {
@@ -44,7 +37,6 @@ export const updateEquipmentInDatabase = async (
     }
   }
 
-  console.log('Final equipment data after validation:', equipmentData);
 
   // For updates, we only filter by equipment ID since RLS policies will handle access control
   // Admin users can update any equipment, regular users can only update their own
@@ -69,6 +61,5 @@ export const updateEquipmentInDatabase = async (
     throw new Error('No data returned from update operation');
   }
 
-  console.log('Equipment updated successfully:', data);
   return data;
 };
